refactor: extract shared RevealMap widget from Triangulation
First step toward showing the post-round reveal map in every game mode
(Daily Scramble and the flight modes come next): the world-map reveal —
equirectangular map, gold answer star, clue-anchor rings, wrong-guess
dots with connector lines, pinch-zoom with constant-size markers,
legend and gesture hint — now lives in core/widgets/reveal_map.dart.
Triangulation's round-result view uses it; its private painter and
per-screen zoom plumbing are deleted. Behaviour is unchanged.
